Move animation logic to its own animation.rs file
